# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger",
    "title": "Recipes API Swagger",
    "contact": {},
    "version": "1.0"
  },
  "host": "localhost:8080",
  "paths": {
    "/login": {
      "post": {
        "description": "Log in with email/username and password to get access and refresh tokens",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "parameters": [
          {
            "description": "Login request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.LoginResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Incorrect username or password",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to generate tokens",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/recipes": {
      "get": {
        "description": "Get multiple recipe previews with query parameters",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Recipes"
        ],
        "summary": "Get Recipes (preview)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "offset",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "author",
            "in": "query",
            "type": "string"
          },
          {
            "name": "title",
            "in": "query",
            "type": "string"
          },
          {
            "name": "preparation_time",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "total_time",
            "in": "query",
            "type": "integer"
          },
          {
            "name": "ingredients",
            "in": "query",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "csv"
          },
          {
            "name": "kind",
            "in": "query",
            "type": "string",
            "enum": [
              "snack",
              "starter",
              "dish",
              "side-dish",
              "sauce",
              "dessert",
              "drink",
              "plate"
            ]
          },
          {
            "name": "locale",
            "in": "query",
            "type": "string",
            "description": "One BCP 47 locale. Missing or stale stored translations are refreshed; failures fall back per recipe to canonical."
          },
          {
            "name": "search_locale",
            "in": "query",
            "type": "string",
            "description": "Optional BCP 47 locale used for title and ingredient substring matching. When omitted, canonical recipe fields are searched."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GetRecipesResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to get the recipes",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Create a complete recipe. Send application/json without new pictures, or multipart/form-data with a recipe JSON field and repeated pictures file fields.",
        "consumes": [
          "application/json",
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Recipes"
        ],
        "summary": "Create Recipe",
        "parameters": [
          {
            "description": "Recipe information",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.CreateRecipe"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "type": "string",
            "description": "Optional retry key retained for 24 hours"
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to create the recipe",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.Recipe"
            }
          },
          "413": {
            "description": "Request too large"
          },
          "415": {
            "description": "Unsupported content type"
          },
          "422": {
            "description": "Incomplete recipe or invalid picture",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/refresh": {
      "post": {
        "description": "Get a new access token by providing a valid refresh token",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Get a new access token",
        "parameters": [
          {
            "description": "Refresh request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RefreshRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.RefreshResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: get the corresponding user / generate a new token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/register": {
      "post": {
        "description": "Create an account with needed information",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Auth"
        ],
        "summary": "Register",
        "parameters": [
          {
            "description": "Register request",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.RegisterRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created",
            "schema": {
              "$ref": "#/definitions/models.MessageResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "406": {
            "description": "Password not strong enough / Invalid mail / Invalid username",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "409": {
            "description": "User already exists",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: generate username / check if username was banned / create user / create jwt / send mail",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get users with filters",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get users",
        "parameters": [
          {
            "type": "integer",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "offset",
            "in": "query"
          },
          {
            "type": "string",
            "name": "username",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.GetUsersResponse"
            }
          },
          "400": {
            "description": "Couldn't read query parameters",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to get users",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get self",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get self",
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/models.UserMe"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to get self",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete self",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete self",
        "parameters": [
          {
            "description": "User information to update",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/models.UserMe"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to delete self / delete picture",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update self",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update self",
        "parameters": [
          {
            "description": "User information to update",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UpdateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/models.UserMe"
            }
          },
          "400": {
            "description": "Bad request",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired jwt",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "406": {
            "description": "Username too long / invalid email / insecure password",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "409": {
            "description": "User already exists",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to update self",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/users/me/picture": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update your profile picture, must be under 8MB",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update self picture",
        "parameters": [
          {
            "type": "file",
            "description": "Picture, formData must be named 'file'",
            "name": "request",
            "in": "formData",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UpdatePictureResponse"
            }
          },
          "400": {
            "description": "Empty formData / not named correctly",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "406": {
            "description": "Wrong format (not png / jpeg / jpg) / File too heavy",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: Update user, save the image",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete your profile picture",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete own picture",
        "responses": {
          "200": {
            "description": "Picture deleted",
            "schema": {
              "$ref": "#/definitions/models.MessageResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "Image not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "422": {
            "description": "No picture for this user",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to: Update user / remove the picture",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Get another user by its id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Get user by id",
        "parameters": [
          {
            "type": "string",
            "description": "User id to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UserView"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "delete": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Delete a user. Must be admin",
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Delete user",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5",
            "description": "User id to delete",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "User deleted",
            "schema": {
              "$ref": "#/definitions/models.UserDB"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Unknown / Insufficient permissions",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "500": {
            "description": "Failed to delete the user / delete associated picture / delete parameters",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      },
      "patch": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "Update another user by its id. Must be admin",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "Users"
        ],
        "summary": "Update user",
        "parameters": [
          {
            "type": "string",
            "example": "6567bd0a84f663dbe91176f5",
            "description": "User id to update",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "Infos to update",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/models.UserDB"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/models.UserDB"
            }
          },
          "400": {
            "description": "Bad request / Username too long",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "401": {
            "description": "Invalid or expired token",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "403": {
            "description": "Unknown / Insufficient permissions",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "404": {
            "description": "User not found",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          },
          "409": {
            "description": "Username or email already taken",
            "schema": {
              "$ref": "#/definitions/models.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "models.CreateRecipe": {
      "type": "object",
      "required": [
        "title",
        "quantity",
        "kind",
        "ingredients",
        "steps"
      ],
      "properties": {
        "title": {
          "type": "string",
          "maxLength": 120
        },
        "description": {
          "type": "string",
          "maxLength": 5000
        },
        "quantity": {
          "type": "integer",
          "minimum": 1
        },
        "kind": {
          "$ref": "#/definitions/models.RecipeKind"
        },
        "preparation_time": {
          "type": "integer",
          "minimum": 0
        },
        "cooking_time": {
          "type": "integer",
          "minimum": 0
        },
        "resting_time": {
          "type": "integer",
          "minimum": 0
        },
        "ingredients": {
          "type": "array",
          "minItems": 1,
          "items": {
            "$ref": "#/definitions/models.Ingredient"
          }
        },
        "steps": {
          "type": "array",
          "minItems": 1,
          "items": {
            "$ref": "#/definitions/models.Step"
          }
        },
        "locale": {
          "type": "string",
          "description": "Canonical BCP 47 locale; detected when omitted"
        }
      }
    },
    "models.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "string",
          "description": "Stable machine-readable error code"
        },
        "request_id": {
          "type": "string",
          "description": "Request identifier, also returned in X-Request-ID"
        }
      }
    },
    "models.GetRecipesResponse": {
      "type": "object",
      "properties": {
        "items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.RecipePreview"
          }
        },
        "length": {
          "type": "integer"
        }
      }
    },
...
    "models.LoginRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "models.LoginResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "expires_at": {
          "type": "string"
        },
        "expires_in": {
          "type": "integer"
        },
        "refresh_token": {
          "type": "string"
        },
        "refresh_token_expires_at": {
          "type": "string"
        },
        "refresh_token_expires_in": {
          "type": "integer"
        }
      }
    },
    "models.MessageResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
...
    "models.RefreshRequest": {
      "type": "object",
      "properties": {
        "refresh_token": {
          "type": "string"
        }
      }
    },
    "models.RefreshResponse": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "expires_at": {
          "type": "string"
        },
        "expires_in": {
          "type": "integer"
        }
      }
    },
    "models.RegisterRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "username": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "basePath": "/api/v1"
}